refactor: remove magical numbers and add global variables file
