+++ async-race/src/modules/data/global-var.ts
@@ -1,0 +1,2 @@
+export const MAIN_URL = "http://127.0.0.1:3000";
+export const CARS_PER_PAGE = 7;
